common: introduce GHOST filter
Since in the last time more and more unserious APRS hardware was brought
to market which do repeat everything as it is (without modify path or
add own call or doing something else with heard data, ...) and therefore
produce real shit on RF we take a try to filter them out.

In few cases they have a fingerprint due to their AX25 iplementation,
which is copied from one buggy hardware to another and with this
fingerprint it is at least theoretically possible to filter the shit
out.

Signed-off-by: Hannes Schmelzer <oe5hpm@oevsv.at>