cleanup: X11 header files
First we drop the X11 directory with copies of some (unknown version)
X11 header files. That was no problem in the past because X11 is quite
old and stable.

In future we do:
- on x86 or native arm build we use the toolchains header (mostly
  located in /usr/include)
- on cross compile targets we provide those header files within the
  lib_xxx folder

We also have to replace the "Xutil.h, Xlib.h, X.h" provided by libxds,
since this files are living also within the X11 project itself - so this
is very confusing and unreadable.
Finally we replace them with the bare minimum needs (to stay compatible)
with xosic.h which does include X11/... files according to the needs.

An at last we drop all really not needed headerfiles:
- cursorfont.h

Signed-off-by: Hannes Schmelzer <oe5hpm@oevsv.at>